refactor(signal): move BreakError and ErrorValue into the signal module
Third step of #1034, groundwork for typing the JIT error channel.
BreakError, ErrorValue, its ERROR_PAYLOAD thread-local, and
take_error_payload move verbatim from eval.rs into src/signal.rs
alongside HaltSignal and PathResultSignal, so every control-flow
signal type lives in one module. eval.rs imports them; no call-site
or behavior changes.

The JIT channel typing (next step) needs these types visible from
jit.rs to convert between anyhow errors and the channel payload
without stringifying.

Refs #1034
